Trying to fix Issue #1
The text files are getting deleted on line 18, 21, and 24 when the StreamWriter function is called. Even though it is only creating an instance of that method, it seems to be running it and replacing the text as empty. Creating a new account, getting passwords, listing the accounts, and changing the account information seem to be functioning.