Fixed Issue #1
I commented out all instances of the StreamWriter or StreamReader method, and replaced them with System.IO.File.ReadAllLines and System.IO.File.WriteAllLines. This was achieved by creating arrays that were one index larger and shorter for adding and deleting an account, respectively, storing the old information in the new array, writing the new array to the text file, then reading the text file into a new instance of the old array that is used by the application. I also added a RefreshArrays(); method that rereads the information from the text file at the end of any button_click method.