Files Encrypt/Decrypt, but can't read on next open
The files successfully encrypt the data, but the decryption for some reason does not work after the application has closed. It simply tries to read the string from the data file before it is decrypted. Perhaps creating a new file as a temporary storage space to read from then dispose of could be of assistance...