Issue #3 Fixed
Files are now encrypted using File.Encrypt and File.Decrypt. Methods to do this in a more efficient way are present, but unused. They used FileStream and it keeps deleting the text inside the files. Again. Currently, the encryption is only good for the computer with multiple users, as the encryption stops other accounts from being able to access these files.